feat(cluster): partition state propagation
ClusterState now has a config field with DesiredPartition count which
tells the leader how many partitions should be present in the cluster.
Controller contains the logic for for managing cluster state.
For now DesiredPartition count is set manually in the code. In the
future we will need to expand Command API to process modifications to
the cluster configuration.

Cluster changes are always two step process. First the Leader writes a
change to the raft. Second the node that needs to act on that change
notifies the leader that it has been processed. These are mostly done by
state changes.
Monitoring of the partition cluster is still in a TODO state.