feat(cluster): make sure that the node can restart and continue
Nodes can now restart almost properly. We still need to work on updating
the state of the cluster but engine is restarted and process instances
that were in progress should continue processing.
Added a lot more tracing to fill out blank spots and help with
debugging.
Fixed issues with persistence for parent process instance.